Add .1" headers and set connector placement
Lots of changes here to add a couple of 0.1" standard headers at the
edges of the board. Most of the Grove connectors have been moved back to
the upright version, except for the board edge connectors which remain
SMD 90 degree connectors. All of the level shifted signals are available
via both Grove connectors and the 0.1" headers (except SPI which is only
on the header).

Due to the major changes in this layout, all of the traces have been
deleted and need to be re-routed.

Signed-off-by: Grant Likely <grant.likely@linaro.org>