Fix part references on SW1 and SW2
KiCad uses the part reference text to figure out which footprint to
update when reading the netlist. This requires the reference field to
exactly match between the schematic and the layout. If the name gets
changed in the layout, then kicad gets very confused. Change the part
reference back to SW1 and SW2, and use the "Value" field to indicate the
power domain.

Signed-off-by: Grant Likely <grant.likely@linaro.org>