Line up placement of 0.1" headers and fix SPI0_SS signal
Make sure the 0.1" headers are precisely lined up with the mounting
holes, and adjust the traces to match. At the same time add dimension
markings so that others can make use of it.

At the same time, fix the SPI0_SS_B label in the schematic. It is was
incorrect on the connector and so the signal was not wired up. Fixed in
schematic, but it still needs to be fixed in the layout.

Signed-off-by: Grant Likely <grant.likely@linaro.org>